move workflow definition modal to common
