+++ packages/frinx-workflow-ui/src/common/modals/diagram-modal.tsx
@@ -1,5 +1,5 @@
 import React, { VoidFunctionComponent } from 'react';
-import WorkflowDia from '../../../executed-workflow-detail/WorkflowDia/WorkflowDia';
+import WorkflowDia from '@frinx/workflow-ui/src/pages/executed-workflow-detail/WorkflowDia/WorkflowDia';
 import {
   Button,
   Modal,
